Fix dictionary access of a string
Products is listed as a list in the arguments for sync_plan(), and is
passed to find_products as a list.  find_products() then tries to
reference the name key of each of the products, causing a traceback
that raises a separate traceback when it tries to call
module.fail_json().